refactor: split layout/page_read.go — level and value decoding into separate files
