+++ layout/page_read.go
@@ -5,7 +5,6 @@ import (
 	"context"
 	"fmt"
 	"io"
-	"math/bits"
 
 	"github.com/apache/thrift/lib/go/thrift"
 
@@ -103,154 +102,6 @@ func ReadPageRawData(thriftReader *thrift.TBufferedTransport, schemaHandler *sch
 	schema := schemaHandler.SchemaElements[pathIndex]
 	page.Schema = schema
 	return page, nil
-}
-
-// extractV2LevelBuffers extracts the level data from a V2 data page and reassembles it into
-// a contiguous buffer with level-length prefixes followed by value data.
-func (p *Page) extractV2LevelBuffers() ([]byte, error) {
-	dll := p.Header.DataPageHeaderV2.GetDefinitionLevelsByteLength()
-	rll := p.Header.DataPageHeaderV2.GetRepetitionLevelsByteLength()
-
-	if dll < 0 || rll < 0 {
-		return nil, fmt.Errorf("GetRLDLFromRawData: invalid level byte lengths (dll=%d, rll=%d)", dll, rll)
-	}
-	rawDataLen := int32(len(p.RawData))
-	if dll+rll > rawDataLen {
-		return nil, fmt.Errorf("GetRLDLFromRawData: level byte lengths exceed raw data size (dll=%d + rll=%d > %d)", dll, rll, rawDataLen)
-	}
-
-	bytesReader := bytes.NewReader(p.RawData)
-	repetitionLevelsBuf, definitionLevelsBuf := make([]byte, rll), make([]byte, dll)
-	dataBuf := make([]byte, len(p.RawData)-int(rll)-int(dll))
-	if _, err := bytesReader.Read(repetitionLevelsBuf); err != nil {
-		return nil, err
-	}
-	if _, err := bytesReader.Read(definitionLevelsBuf); err != nil {
-		return nil, err
-	}
-	if _, err := bytesReader.Read(dataBuf); err != nil {
-		return nil, err
-	}
-
-	buf := make([]byte, 0)
-	if rll > 0 {
-		tmpBuf, err := encoding.WritePlainINT32([]any{int32(rll)})
-		if err != nil {
-			return nil, err
-		}
-		buf = append(buf, tmpBuf...)
-		buf = append(buf, repetitionLevelsBuf...)
-	}
-	if dll > 0 {
-		tmpBuf, err := encoding.WritePlainINT32([]any{int32(dll)})
-		if err != nil {
-			return nil, err
-		}
-		buf = append(buf, tmpBuf...)
-		buf = append(buf, definitionLevelsBuf...)
-	}
-	buf = append(buf, dataBuf...)
-	return buf, nil
-}
-
-// readLevelValues reads either repetition or definition level values from a reader.
-func readLevelValues(bytesReader *bytes.Reader, maxLevel int32, numValues uint64) ([]any, error) {
-	if maxLevel > 0 {
-		bitWidth := uint64(bits.Len32(uint32(maxLevel)))
-		levels, err := ReadDataPageValues(bytesReader, parquet.Encoding_RLE, parquet.Type_INT64, -1, numValues, bitWidth)
-		if err != nil {
-			return nil, err
-		}
-		if uint64(len(levels)) > numValues {
-			levels = levels[:numValues]
-		}
-		return levels, nil
-	}
-	levels := make([]any, numValues)
-	for i := range len(levels) {
-		levels[i] = int64(0)
-	}
-	return levels, nil
-}
-
-// Get RepetitionLevels and Definitions from RawData
-func (p *Page) GetRLDLFromRawData(schemaHandler *schema.SchemaHandler) (int64, int64, error) {
-	var buf []byte
-	var err error
-
-	if p.Header.GetType() == parquet.PageType_DATA_PAGE_V2 {
-		buf, err = p.extractV2LevelBuffers()
-		if err != nil {
-			return 0, 0, err
-		}
-	} else if p.CompressType != parquet.CompressionCodec_UNCOMPRESSED {
-		buf, err = resolveCompressor(p.compressor).UncompressWithExpectedSize(p.RawData, p.CompressType, int64(p.Header.GetUncompressedPageSize()))
-		if err != nil {
-			return 0, 0, fmt.Errorf("uncompress data page: %w", err)
-		}
-	} else {
-		buf = p.RawData
-	}
-
-	switch p.Header.GetType() {
-	case parquet.PageType_DATA_PAGE, parquet.PageType_DATA_PAGE_V2:
-		return p.decodeDataPageLevels(buf, schemaHandler)
-	case parquet.PageType_DICTIONARY_PAGE:
-		table := new(Table)
-		table.Path = p.Path
-		p.DataTable = table
-		p.RawData = buf
-		return 0, 0, nil
-	default:
-		return 0, 0, fmt.Errorf("unsupported page type: %v", p.Header.GetType())
-	}
-}
-
-func (p *Page) decodeDataPageLevels(buf []byte, schemaHandler *schema.SchemaHandler) (int64, int64, error) {
-	var numValues uint64
-	if p.Header.GetType() == parquet.PageType_DATA_PAGE {
-		numValues = uint64(p.Header.DataPageHeader.GetNumValues())
-	} else {
-		numValues = uint64(p.Header.DataPageHeaderV2.GetNumValues())
-	}
-
-	maxDefinitionLevel, _ := schemaHandler.MaxDefinitionLevel(p.Path)
-	maxRepetitionLevel, _ := schemaHandler.MaxRepetitionLevel(p.Path)
-
-	bytesReader := bytes.NewReader(buf)
-	repetitionLevels, err := readLevelValues(bytesReader, maxRepetitionLevel, numValues)
-	if err != nil {
-		return 0, 0, err
-	}
-	definitionLevels, err := readLevelValues(bytesReader, maxDefinitionLevel, numValues)
-	if err != nil {
-		return 0, 0, err
-	}
-
-	table := new(Table)
-	table.Path = p.Path
-	name := common.PathToStr(p.Path)
-	table.RepetitionType = schemaHandler.SchemaElements[schemaHandler.MapIndex[name]].GetRepetitionType()
-	table.MaxRepetitionLevel = maxRepetitionLevel
-	table.MaxDefinitionLevel = maxDefinitionLevel
-	table.Values = make([]any, len(definitionLevels))
-	table.RepetitionLevels = make([]int32, len(definitionLevels))
-	table.DefinitionLevels = make([]int32, len(definitionLevels))
-
-	numRows := int64(0)
-	for i := range len(definitionLevels) {
-		dl, _ := definitionLevels[i].(int64)
-		rl, _ := repetitionLevels[i].(int64)
-		table.RepetitionLevels[i] = int32(rl)
-		table.DefinitionLevels[i] = int32(dl)
-		if table.RepetitionLevels[i] == 0 {
-			numRows++
-		}
-	}
-	p.DataTable = table
-	p.RawData = buf[len(buf)-bytesReader.Len():]
-
-	return int64(numValues), numRows, nil
 }
 
 // Get values from raw data
@@ -358,136 +209,6 @@ func ReadPageHeader(thriftReader *thrift.TBufferedTransport) (*parquet.PageHeade
 	pageHeader := parquet.NewPageHeader()
 	err := pageHeader.Read(context.TODO(), protocol)
 	return pageHeader, err
-}
-
-// convertRLEValuesForType converts int64 RLE values to the target data type (int32 or bool).
-func convertRLEValuesForType(values []any, dataType parquet.Type) {
-	switch dataType {
-	case parquet.Type_INT32:
-		for i := range values {
-			values[i] = int32(values[i].(int64))
-		}
-	case parquet.Type_BOOLEAN:
-		for i := range values {
-			values[i] = values[i].(int64) > 0
-		}
-	}
-}
-
-// readByteStreamSplit reads values using the BYTE_STREAM_SPLIT encoding for the given data type.
-func readByteStreamSplit(bytesReader *bytes.Reader, dataType parquet.Type, cnt, bitWidth uint64) ([]any, error) {
-	switch dataType {
-	case parquet.Type_FLOAT:
-		return encoding.ReadByteStreamSplitFloat32(bytesReader, cnt)
-	case parquet.Type_DOUBLE:
-		return encoding.ReadByteStreamSplitFloat64(bytesReader, cnt)
-	case parquet.Type_INT32:
-		return encoding.ReadByteStreamSplitINT32(bytesReader, cnt)
-	case parquet.Type_INT64:
-		return encoding.ReadByteStreamSplitINT64(bytesReader, cnt)
-	case parquet.Type_FIXED_LEN_BYTE_ARRAY:
-		return encoding.ReadByteStreamSplitFixedLenByteArray(bytesReader, cnt, bitWidth)
-	default:
-		return nil, fmt.Errorf("the encoding method BYTE_STREAM_SPLIT is only supported for FLOAT, DOUBLE, INT32, INT64, FIXED_LEN_BYTE_ARRAY, got %v", dataType)
-	}
-}
-
-// defaultBitWidth returns the default bitWidth for a data type when not specified.
-func defaultBitWidth(dataType parquet.Type) uint64 {
-	switch dataType {
-	case parquet.Type_BOOLEAN:
-		return 1
-	case parquet.Type_INT32:
-		return 32
-	case parquet.Type_INT64:
-		return 64
-	default:
-		return 0
-	}
-}
-
-func readDeltaBinaryPacked(bytesReader *bytes.Reader, dataType parquet.Type) ([]any, error) {
-	switch dataType {
-	case parquet.Type_INT32:
-		return encoding.ReadDeltaBinaryPackedINT32(bytesReader)
-	case parquet.Type_INT64:
-		return encoding.ReadDeltaBinaryPackedINT64(bytesReader)
-	default:
-		return nil, fmt.Errorf("the encoding method DELTA_BINARY_PACKED can only be used with int32 and int64 types, got %v", dataType)
-	}
-}
-
-func readDeltaByteArrayValues(values []any, err error, dataType parquet.Type, cnt uint64) ([]any, error) {
-	if err != nil {
-		return nil, err
-	}
-	if dataType == parquet.Type_FIXED_LEN_BYTE_ARRAY {
-		for i := range values {
-			values[i] = values[i].(string)
-		}
-	}
-	return values[:cnt], nil
-}
-
-// Read data page values
-func ReadDataPageValues(bytesReader *bytes.Reader, encodingMethod parquet.Encoding, dataType parquet.Type, convertedType parquet.ConvertedType, cnt, bitWidth uint64) ([]any, error) {
-	var res []any
-	if cnt <= 0 {
-		return res, nil
-	}
-
-	if bitWidth == 0 {
-		bitWidth = defaultBitWidth(dataType)
-	}
-
-	switch encodingMethod {
-	case parquet.Encoding_PLAIN:
-		return encoding.ReadPlain(bytesReader, dataType, cnt, bitWidth)
-	case parquet.Encoding_PLAIN_DICTIONARY, parquet.Encoding_RLE_DICTIONARY:
-		b, err := bytesReader.ReadByte()
-		if err != nil {
-			return res, err
-		}
-		bitWidth = uint64(b)
-
-		buf, err := encoding.ReadRLEBitPackedHybrid(bytesReader, bitWidth, uint64(bytesReader.Len()))
-		if err != nil {
-			return res, err
-		}
-		if uint64(len(buf)) < cnt {
-			return res, fmt.Errorf("expected %d values but got %d from RLE/bit-packed hybrid decoder", cnt, len(buf))
-		}
-		return buf[:cnt], err
-	case parquet.Encoding_RLE:
-		values, err := encoding.ReadRLEBitPackedHybrid(bytesReader, bitWidth, 0)
-		if err != nil {
-			return res, err
-		}
-		if uint64(len(values)) < cnt {
-			return res, fmt.Errorf("expected %d values but got %d from RLE/bit-packed hybrid decoder", cnt, len(values))
-		}
-		convertRLEValuesForType(values, dataType)
-		return values[:cnt], nil
-	case parquet.Encoding_BIT_PACKED:
-		values, err := encoding.ReadBitPackedCount(bytesReader, cnt, bitWidth)
-		if err != nil {
-			return res, err
-		}
-		convertRLEValuesForType(values, dataType)
-		return values, nil
-	case parquet.Encoding_DELTA_BINARY_PACKED:
-		return readDeltaBinaryPacked(bytesReader, dataType)
-	case parquet.Encoding_DELTA_LENGTH_BYTE_ARRAY:
-		values, err := encoding.ReadDeltaLengthByteArray(bytesReader)
-		return readDeltaByteArrayValues(values, err, dataType, cnt)
-	case parquet.Encoding_DELTA_BYTE_ARRAY:
-		values, err := encoding.ReadDeltaByteArray(bytesReader)
-		return readDeltaByteArrayValues(values, err, dataType, cnt)
-	case parquet.Encoding_BYTE_STREAM_SPLIT:
-		return readByteStreamSplit(bytesReader, dataType, cnt, bitWidth)
-	default:
-		return res, fmt.Errorf("unknown Encoding method: %v", encodingMethod)
-	}
 }
 
 // readPageV2Data reads a DATA_PAGE_V2 from the reader, decompresses if needed, and reassembles with level prefixes.
